HTTP-webserver and mDNS added
An HTTP-webserver on port 80 was added. There, correction values can be applied. The ESP can now be reached at windsensor.local from within the own network. 
NOTE: This does not work on Android devices.
Also, the blinking of the IP can be turned on and of via a button on the webserver